rename test.rs to test_gateway.rs
