+++ contracts/axelar-gateway/src/lib.rs
@@ -15,7 +15,7 @@ compile_error!("'testutils' feature is not supported on 'wasm' target");
 pub mod testutils;
 
 #[cfg(test)]
-mod test;
+mod test_gateway;
 
 #[cfg(test)]
 mod test_auth;
